Predict Aboveground Biomass Using RSYC models

This function generates aboveground biomass (AGB) predictions using remote sensing-based yield curves (RSYC) for a given species, stand age, and tile. It retrieves the model parameters associated with the specified tile and species and applies a four-parameter Chapman-Richards growth function.

Usage

predict_rsyc(tile_id, age, species)

Arguments

tile_id

A character string specifying the tile ID. Must match a tile for which RSYC models are available.

age

A numeric vector of stand ages (in years) at which to predict biomass. All values must be positive.

species

A character string specifying the species code (e.g., "PICE.MAR" for black spruce).

Value

A numeric vector of predicted AGB values (e.g., in tonnes per hectare) corresponding to the input ages.

Details

The RSYC models were developed using data from 0 to 150 years. If any ages provided exceed 150, a warning is issued because predictions represent extrapolations beyond the model's calibration range and may be unreliable.

Examples

predict_rsyc(tile_id = "J11", age = 1:100, species = "PICE.MAR")
#> Error in loadNamespace(x): there is no package called ‘lazyeval’